Moved expirmental map components into folder for organizational purposes
